Enable timeout by default

+++ src/libraries/JANA/CLI/JMain.cc
@@ -122,7 +122,7 @@ int Execute(JApplication* app, UserOptions &options) {
         // Run JANA in normal mode
         app->Run();
     }
-    return app->GetExitCode();
+    return (int) app->GetExitCode();
 }
